clarification on logic for updating a null hand
In the old behavior, if the server had a null hand and the client sent a
blank hand (it can't send null), the server would always claim there was
a change to null, and this spammed the logs.  The new change will only
add the change if it's going from a non-null to a "blank" from the
client (ie the client wants to make it null in the db).